center dto 넘길때 이름아닌 아이디로 정보 넘겨주기, boolean 오류 수정, 이력서 GET 동작 확인

+++ Maeum-Eum/src/main/java/com/five/Maeum_Eum/dto/user/caregiver/resume/response/ResumeResponseDTO.java
@@ -20,7 +20,7 @@ public class ResumeResponseDTO {
     private Boolean isNegotiableTime;
     private Integer wage;
     private Integer expectedSalary;
-    private List<Long> elderRank;
+    private List<Integer> elderRank;
     private List<String> meal;
     private List<String> toileting;
     private List<String> mobility;
